+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Under development
 
+- Profile
+    - Scope: Validate regular expressions.
+
 ## 0.5.10 _(February 8, 2015)_
 
 - Switched to absolute paths instead of full URLs across the interface.